requirements.txt
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
firebase-admin==6.6.0
Flask==3.1.0  # ✅ Mantida apenas a versão mais recente
frozenlist==1.5.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-cloud-firestore==2.20.0
grpcio==1.70.0
grpcio-status==1.70.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
oauthlib==3.2.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pycparser==2.22
pydub==0.25.1
PyJWT==2.10.1
pyparsing==3.2.1
python-telegram-bot[webhooks]==21.1.1
requests==2.32.3
rsa==4.9
SpeechRecognition==3.14.1
twilio==9.4.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
python-dotenv==1.0.1
apscheduler==3.10.4
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2
gunicorn==21.2.0
gevent==24.2.1
greenlet==3.0.3
dateparser==1.2.1
google-auth-oauthlib
gTTS==2.5.4
openai>=1.0.0
unidecode==1.3.8
pandas==2.2.2
openpyxl==3.1.5
